Move tests and model of car into car_information
